requirements.txt
valohai-utils
monai
nibabel
numpy
torch
scikit-learn
matplotlib
tqdm